Consolidate backend/vault and vault/vault
